% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{analysis_config}
\alias{analysis_config}
\title{Create analysis configuration}
\usage{
analysis_config(
  radii = seq(50, 5000, by = 50),
  bandwidth = 100,
  weight_type = "ring",
  annular_width = 50
)
}
\arguments{
\item{radii}{Vector of radii to evaluate}

\item{bandwidth}{Ring bandwidth}

\item{weight_type}{Weight type: "ring", "gaussian", or "annular"}

\item{annular_width}{Annular width}
}
\value{
Analysis configuration list
}
\description{
Create analysis configuration
}
