% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sc.R
\name{as_data_sc}
\alias{as_data_sc}
\title{Convert Seurat Object to SCData}
\usage{
as_data_sc(seurat_obj, assay = "RNA", cell_type_col = "cell_type")
}
\arguments{
\item{seurat_obj}{Seurat object with spatial data}

\item{assay}{Assay name (default: "RNA")}

\item{cell_type_col}{Column in metadata for cell type (default: "cell_type")}
}
\value{
SCData list
}
\description{
Convert a Seurat object with spatial data to SCData format.
}
