% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calculate_lambda}
\alias{calculate_lambda}
\title{Calculate characteristic diffusion length}
\usage{
calculate_lambda(diffusion_config, n_receivers, p_r = 1)
}
\arguments{
\item{diffusion_config}{Diffusion configuration}

\item{n_receivers}{Receiver density}

\item{p_r}{Probability of responding}
}
\value{
Lambda (characteristic diffusion length)
}
\description{
Calculate characteristic diffusion length
}
