% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{celltype_config}
\alias{celltype_config}
\title{Create cell type configuration}
\usage{
celltype_config(
  n_active_senders = 20,
  n_silent_senders = 0,
  receiver_fractions = c(0.1, 0.2, 0.3, 0.5)
)
}
\arguments{
\item{n_active_senders}{Number of active sender cells}

\item{n_silent_senders}{Number of silent sender cells}

\item{receiver_fractions}{Vector of receiver fractions to simulate}
}
\value{
Cell type configuration list
}
\description{
Create cell type configuration
}
