% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse10m_analysis.R
\name{compute_delta_ind}
\alias{compute_delta_ind}
\title{Compute delta I_ND between active and control conditions}
\usage{
compute_delta_ind(ind_active, ind_control)
}
\arguments{
\item{ind_active}{I_ND results for active condition}

\item{ind_control}{I_ND results for control condition}
}
\value{
Data frame with delta I_ND
}
\description{
Compute delta I_ND between active and control conditions
}
