% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_metric_batch}
\alias{compute_metric_batch}
\title{Batch Compute Metrics for All Genes}
\usage{
compute_metric_batch(z_f, lag_G, metric = c("moran", "ind"))
}
\arguments{
\item{z_f}{Numeric vector. Standardized factor expression (length n).}

\item{lag_G}{Numeric matrix. Spatial lag matrix (n x n_genes), where each
column is W * z_g for gene g.}

\item{metric}{Character. Either "moran" or "ind".}
}
\value{
Numeric vector of metric values (length n_genes).
}
\description{
Efficiently computes Moran's I or I_ND for one factor against all genes
using pre-computed spatial lag matrix.
}
\examples{
n <- 100
n_genes <- 50
z_f <- rnorm(n)
lag_G <- matrix(rnorm(n * n_genes), n, n_genes)
compute_metric_batch(z_f, lag_G, "ind")

}
\seealso{
Other core: 
\code{\link{compute_ind_from_lag}()},
\code{\link{compute_moran_from_lag}()},
\code{\link{compute_spatial_lag}()},
\code{\link{compute_spatial_lag_batch}()}
}
\concept{core}
