% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_moran_from_lag}
\alias{compute_moran_from_lag}
\title{Compute Bivariate Moran's I from Pre-computed Spatial Lag}
\usage{
compute_moran_from_lag(z_f, lag_g)
}
\arguments{
\item{z_f}{Numeric vector. Standardized factor expression (length n).}

\item{lag_g}{Numeric vector. Spatial lag of gene expression (W * z_g),
same length as z_f.}
}
\value{
Scalar Moran's I value (unbounded).
}
\description{
Calculates Moran's I statistic given a factor expression vector and
pre-computed spatial lag of gene expression.
}
\details{
The bivariate Moran's I is computed as:
\deqn{I = \frac{z_f' \cdot lag_g}{n}}

where:
\itemize{
\item \code{z_f} is the standardized factor expression
\item \code{lag_g} is the pre-computed spatial lag (W * z_g)
\item \code{n} is the number of observations
}
}
\examples{
# Simple example
z_f <- c(1, -1, 1, -1)
lag_g <- c(0.5, -0.5, 0.5, -0.5)
compute_moran_from_lag(z_f, lag_g)

}
\seealso{
Other core: 
\code{\link{compute_ind_from_lag}()},
\code{\link{compute_metric_batch}()},
\code{\link{compute_spatial_lag}()},
\code{\link{compute_spatial_lag_batch}()}
}
\concept{core}
