% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_moran_from_lag_cpp}
\alias{compute_moran_from_lag_cpp}
\title{Compute Bivariate Moran's I from Pre-computed Spatial Lag}
\usage{
compute_moran_from_lag_cpp(z_f, lag_g)
}
\arguments{
\item{z_f}{Standardized factor expression vector (length n)}

\item{lag_g}{Pre-computed spatial lag of gene (W * z_g), same length as z_f}
}
\value{
Scalar Moran's I value (unbounded)
}
\description{
Calculates Moran's I statistic using the formula:
I = z_f' * lag_g / n
}
\details{
where lag_g = W * z_g has been pre-computed.

Moran's I measures spatial autocorrelation. Positive values indicate
spatial clustering (similar values near each other), negative values
indicate dispersion (dissimilar values near each other).
}
