% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_spatial_lag}
\alias{compute_spatial_lag}
\title{Compute Spatial Lag (W * z)}
\usage{
compute_spatial_lag(W, z)
}
\arguments{
\item{W}{Sparse weight matrix (n x m), typically row-normalized.}

\item{z}{Numeric vector. Expression values (length m).}
}
\value{
Numeric vector. Spatial lag (length n).
}
\description{
Computes the spatial lag of an expression vector using a weight matrix.
}
\details{
The spatial lag represents the weighted average of neighboring values:
\deqn{lag_i = \sum_j w_{ij} z_j}
}
\examples{
library(Matrix)
W <- sparseMatrix(i = c(1,1,2,2), j = c(2,3,1,3), x = c(0.5,0.5,0.5,0.5))
z <- c(1, 2, 3)
compute_spatial_lag(W, z)

}
\seealso{
Other core: 
\code{\link{compute_ind_from_lag}()},
\code{\link{compute_metric_batch}()},
\code{\link{compute_moran_from_lag}()},
\code{\link{compute_spatial_lag_batch}()}
}
\concept{core}
