% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_spatial_lag_batch}
\alias{compute_spatial_lag_batch}
\title{Batch Compute Spatial Lags (W * Z)}
\usage{
compute_spatial_lag_batch(W, Z)
}
\arguments{
\item{W}{Sparse weight matrix (n x m), row-normalized.}

\item{Z}{Numeric matrix. Expression matrix (m x n_genes), genes in columns.}
}
\value{
Numeric matrix. Spatial lag matrix (n x n_genes).
}
\description{
Computes spatial lags for all genes at once using matrix multiplication.
}
\examples{
library(Matrix)
n <- 50
m <- 50
n_genes <- 100
W <- rsparsematrix(n, m, density = 0.1)
W <- W / rowSums(abs(W))
Z <- matrix(rnorm(m * n_genes), m, n_genes)
lag_G <- compute_spatial_lag_batch(W, Z)

}
\seealso{
Other core: 
\code{\link{compute_ind_from_lag}()},
\code{\link{compute_metric_batch}()},
\code{\link{compute_moran_from_lag}()},
\code{\link{compute_spatial_lag}()}
}
\concept{core}
