% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{diffusion_config}
\alias{diffusion_config}
\title{Create diffusion configuration}
\usage{
diffusion_config(D = 100, k_max = 10, Kd = 1, secretion_rate = 1)
}
\arguments{
\item{D}{Diffusion coefficient (um^2/s)}

\item{k_max}{Maximum uptake rate}

\item{Kd}{Dissociation constant}

\item{secretion_rate}{Expression to concentration scaling}
}
\value{
Diffusion configuration list
}
\description{
Create diffusion configuration
}
