% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{.build_weight_matrix}
\alias{.build_weight_matrix}
\title{Build weight matrix based on weight type}
\usage{
.build_weight_matrix(dists, radius, analysis_config)
}
\description{
Build weight matrix based on weight type
}
\keyword{internal}
