% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{.generate_random_positions}
\alias{.generate_random_positions}
\title{Generate random positions with minimum separation}
\usage{
.generate_random_positions(config, domain_config)
}
\description{
Generate random positions with minimum separation
}
\keyword{internal}
