% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomewide_analysis.R
\name{.get_cell_type_pairs_internal}
\alias{.get_cell_type_pairs_internal}
\title{Get Cell Type Pairs for Analysis}
\usage{
.get_cell_type_pairs_internal(data, pairs, min_cells, verbose)
}
\description{
Get Cell Type Pairs for Analysis
}
\keyword{internal}
