% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_pair_analysis.R
\name{.init_hdf5_output}
\alias{.init_hdf5_output}
\title{Initialize HDF5 Output File}
\usage{
.init_hdf5_output(h5_path, n_pairs, n_genes, n_radii, pairs, radii, gene_names)
}
\description{
Initialize HDF5 Output File
}
\keyword{internal}
