% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{.save_simulation_results}
\alias{.save_simulation_results}
\title{Save simulation results to files}
\usage{
.save_simulation_results(results, config)
}
\description{
Save simulation results to files
}
\keyword{internal}
