% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomewide_analysis.R
\name{.validate_genomewide_inputs}
\alias{.validate_genomewide_inputs}
\title{Validate Genomewide Analysis Inputs}
\usage{
.validate_genomewide_inputs(data)
}
\description{
Validate Genomewide Analysis Inputs
}
\keyword{internal}
