% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_pair_analysis.R
\name{.write_pair_to_hdf5}
\alias{.write_pair_to_hdf5}
\title{Write Pair Results to HDF5}
\usage{
.write_pair_to_hdf5(h5_path, pair_idx, result)
}
\description{
Write Pair Results to HDF5
}
\keyword{internal}
