% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{expression_config}
\alias{expression_config}
\title{Create expression configuration}
\usage{
expression_config(
  F_basal = 0.1,
  F_high = 10,
  R_basal = 0.1,
  fold_change = 10,
  sigma_f = 0.1,
  sigma_f_basal = 0.1,
  sigma_r = 0.1,
  sigma_r_basal = 0.1
)
}
\arguments{
\item{F_basal}{Basal factor expression}

\item{F_high}{High factor expression for active senders}

\item{R_basal}{Basal response expression}

\item{fold_change}{Response fold change}

\item{sigma_f}{Lognormal sigma for expressing senders}

\item{sigma_f_basal}{Lognormal sigma for non-expressing senders}

\item{sigma_r}{Lognormal sigma for responding receivers}

\item{sigma_r_basal}{Lognormal sigma for non-responding receivers}
}
\value{
Expression configuration list
}
\description{
Create expression configuration
}
