% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.R
\name{filter_in_tissue}
\alias{filter_in_tissue}
\title{Filter Visium Data to In-Tissue Spots}
\usage{
filter_in_tissue(visium_data)
}
\arguments{
\item{visium_data}{A list returned by \code{load_visium_data}.}
}
\value{
A list with the same structure as input, but filtered to in-tissue spots only.
}
\description{
Subset Visium data to include only spots that are within the tissue.
}
\examples{
\dontrun{
visium <- load_visium_data("path/to/spaceranger_output")
visium_tissue <- filter_in_tissue(visium)
}

}
