% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_cell_positions}
\alias{generate_cell_positions}
\title{Generate cell positions in circular domain}
\usage{
generate_cell_positions(config)
}
\arguments{
\item{config}{Domain configuration}
}
\value{
Matrix of cell positions (n_cells x 2)
}
\description{
Generate cell positions in circular domain
}
