% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_factor_expression}
\alias{generate_factor_expression}
\title{Generate stochastic factor expression}
\usage{
generate_factor_expression(
  n_total,
  n_active,
  active_indices,
  expr_config,
  stoch_config
)
}
\arguments{
\item{n_total}{Total number of cells}

\item{n_active}{Number of active senders}

\item{active_indices}{Indices of active sender cells}

\item{expr_config}{Expression configuration}

\item{stoch_config}{Stochastic configuration}
}
\value{
List with factor_expr and expressing_mask
}
\description{
Model:
F_i = S_i * F_high * LogNormal(0, sigma_f^2) + (1-S_i) * F_basal * LogNormal(0, sigma_f_basal^2)
S_i ~ Bernoulli(p_sender_express)
}
