% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse10m_analysis.R
\name{generate_inert_positions}
\alias{generate_inert_positions}
\title{Generate inert cell positions}
\usage{
generate_inert_positions(n, tissue_radius)
}
\arguments{
\item{n}{Number of positions}

\item{tissue_radius}{Tissue radius}
}
\value{
Matrix of positions (n x 2)
}
\description{
Generate inert cell positions
}
