% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse10m_analysis.R
\name{generate_receiver_positions}
\alias{generate_receiver_positions}
\title{Generate receiver positions with distance-dependent mixing}
\usage{
generate_receiver_positions(n, tissue_radius, source_region_radius)
}
\arguments{
\item{n}{Number of positions}

\item{tissue_radius}{Tissue radius}

\item{source_region_radius}{Source region radius}
}
\value{
Matrix of positions (n x 2)
}
\description{
Generate receiver positions with distance-dependent mixing
}
