% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_sender_positions}
\alias{generate_sender_positions}
\title{Generate sender position dictionary}
\usage{
generate_sender_positions(config, domain_config)
}
\arguments{
\item{config}{Position configuration}

\item{domain_config}{Domain configuration}
}
\value{
Named list of position coordinates
}
\description{
Generate sender position dictionary
}
