% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse10m_analysis.R
\name{generate_source_positions}
\alias{generate_source_positions}
\title{Generate positions in source region (central circular area)}
\usage{
generate_source_positions(n, source_region_radius)
}
\arguments{
\item{n}{Number of positions}

\item{source_region_radius}{Radius of source region}
}
\value{
Matrix of positions (n x 2)
}
\description{
Generate positions in source region (central circular area)
}
