% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_weights.R
\name{load_dense_weight_matrix_internal}
\alias{load_dense_weight_matrix_internal}
\title{Load Dense Weight Matrix}
\usage{
load_dense_weight_matrix_internal(file, sep)
}
\description{
Load Dense Weight Matrix
}
\keyword{internal}
