% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_sc.R
\name{load_hdf5_sc}
\alias{load_hdf5_sc}
\title{Load Results from HDF5}
\usage{
load_hdf5_sc(file)
}
\arguments{
\item{file}{Path to HDF5 file}
}
\value{
List with IND_array and metadata
}
\description{
Load batch computation results from HDF5 file.
}
