% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{load_moran_result}
\alias{load_moran_result}
\title{Load Moran's I Results}
\usage{
load_moran_result(input_file)
}
\arguments{
\item{input_file}{Path to input file.}
}
\value{
A symmetric matrix of pairwise Moran's I values.
}
\description{
Load pairwise Moran's I results from a file in lower triangular format.
}
