% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_weights.R
\name{load_sparse_coo_matrix_internal}
\alias{load_sparse_coo_matrix_internal}
\title{Load Sparse COO Matrix}
\usage{
load_sparse_coo_matrix_internal(file, sep, spot_names = NULL)
}
\description{
Load Sparse COO Matrix
}
\keyword{internal}
