% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_visium.R
\name{moran_I}
\alias{moran_I}
\title{Compute Moran's I for a Single Gene Pair}
\usage{
moran_I(x, y, W, normalize = TRUE)
}
\arguments{
\item{x}{Numeric vector of expression values for gene 1.}

\item{y}{Numeric vector of expression values for gene 2.}

\item{W}{Weight matrix (spots x spots).}

\item{normalize}{Whether to z-normalize the input vectors. Default: TRUE.}
}
\value{
A single numeric value (Moran's I).
}
\description{
Compute Moran's I between two specific genes.
}
