% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{normalize_expression}
\alias{normalize_expression}
\title{Normalize expression data}
\usage{
normalize_expression(raw_expr, norm_config)
}
\arguments{
\item{raw_expr}{Raw expression vector}

\item{norm_config}{Normalization configuration}
}
\value{
Normalized expression vector
}
\description{
Normalize expression data
}
