% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairwise_moran_matrix_cpp}
\alias{pairwise_moran_matrix_cpp}
\title{Fast Pairwise Moran's I Computation}
\usage{
pairwise_moran_matrix_cpp(data_z, W, S0)
}
\arguments{
\item{data_z}{Z-normalized expression matrix (genes x spots)}

\item{W}{Weight matrix (spots x spots)}

\item{S0}{Sum of all weights}
}
\value{
Moran's I matrix (genes x genes)
}
\description{
Computes pairwise Moran's I for all gene pairs using optimized BLAS
matrix operations.
}
\keyword{internal}
