% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse10m_analysis.R
\name{parse10m_config}
\alias{parse10m_config}
\title{Create Parse10M analysis configuration}
\usage{
parse10m_config(
  h5ad_path = "",
  output_dir = "./parse10m_analysis",
  source_cell_type = "CD4 Memory",
  receiver_cell_type = "CD14 Mono",
  inert_cell_type = "B Naive",
  cytokine_name = "IFN-gamma",
  ligand_gene = "IFNG",
  tissue_radius = 5000,
  source_region_radius = 500,
  sigma_decay = 1500,
  p_cytokine_at_source = 0.9,
  p_cytokine_background = 0.1,
  n_source_cells = 500,
  n_control_sources = 500,
  n_receiver_cells = 15000,
  n_inert_cells = 5000,
  log2fc_threshold = 0.5,
  pval_threshold = 0.05,
  radii = seq(50, 5000, by = 50),
  use_annular = TRUE,
  annular_width = 50,
  response_genes = c("MX1", "IFITM2", "IFITM3", "IFI6", "IFI27", "IFIT1", "IFIT2",
    "IFIT3", "IRF1", "GBP1", "GBP2", "STAT1", "B2M", "HLA-A", "HLA-B", "HLA-C"),
  random_seed = 42
)
}
\arguments{
\item{h5ad_path}{Path to Parse10M h5ad file}

\item{output_dir}{Output directory}

\item{source_cell_type}{Source/sender cell type}

\item{receiver_cell_type}{Receiver cell type}

\item{inert_cell_type}{Inert (non-responsive) cell type}

\item{cytokine_name}{Cytokine name for treatment}

\item{ligand_gene}{Ligand gene name}

\item{tissue_radius}{Tissue radius (um)}

\item{source_region_radius}{Source region radius (um)}

\item{sigma_decay}{Expression decay sigma}

\item{p_cytokine_at_source}{Probability of cytokine-treated cell at source}

\item{p_cytokine_background}{Probability of cytokine-treated cell in background}

\item{n_source_cells}{Number of source cells}

\item{n_control_sources}{Number of control source cells}

\item{n_receiver_cells}{Number of receiver cells}

\item{n_inert_cells}{Number of inert cells}

\item{log2fc_threshold}{Log2 fold change threshold for DE}

\item{pval_threshold}{P-value threshold for DE}

\item{radii}{Vector of radii for I_ND analysis}

\item{use_annular}{Use annular weights}

\item{annular_width}{Annular width}

\item{response_genes}{Known response genes}

\item{random_seed}{Random seed}
}
\value{
Configuration list
}
\description{
Create Parse10M analysis configuration
}
