% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse10m_analysis.R
\name{perform_de_analysis}
\alias{perform_de_analysis}
\title{Perform differential expression analysis between two conditions}
\usage{
perform_de_analysis(
  expr_matrix,
  condition_labels,
  condition1,
  condition2,
  min_pct = 0.1
)
}
\arguments{
\item{expr_matrix}{Expression matrix (genes x cells)}

\item{condition_labels}{Vector of condition labels for each cell}

\item{condition1}{Name of first condition (e.g., "cytokine")}

\item{condition2}{Name of second condition (e.g., "PBS")}

\item{min_pct}{Minimum percentage of cells expressing gene}
}
\value{
Data frame with DE results
}
\description{
Perform differential expression analysis between two conditions
}
