% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation}
\alias{permutation}
\title{Permutation Tests for Spatial Correlation}
\description{
Functions for permutation-based significance testing of Moran's I
and I_ND spatial correlation metrics.
}
