% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{plot_cell_distribution}
\alias{plot_cell_distribution}
\title{Plot cell distribution}
\usage{
plot_cell_distribution(results, output_path = NULL)
}
\arguments{
\item{results}{Simulation results (single receiver fraction)}

\item{output_path}{Output file path (PNG)}
}
\description{
Plot cell distribution
}
