% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse10m_analysis.R
\name{plot_delta_ind_profile}
\alias{plot_delta_ind_profile}
\title{Plot delta I_ND profile}
\usage{
plot_delta_ind_profile(delta_ind, response_genes, output_path = NULL)
}
\arguments{
\item{delta_ind}{Delta I_ND results}

\item{response_genes}{Known response genes}

\item{output_path}{Output file path (PNG)}
}
\description{
Plot delta I_ND profile
}
