% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse10m_analysis.R
\name{plot_parse10m_layout}
\alias{plot_parse10m_layout}
\title{Plot spatial layout}
\usage{
plot_parse10m_layout(results, output_path = NULL)
}
\arguments{
\item{results}{Results from run_parse10m_simulation()}

\item{output_path}{Output file path (PNG)}
}
\description{
Plot spatial layout
}
