% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sc.R
\name{print.SCData}
\alias{print.SCData}
\title{Print Method for SCData}
\usage{
\method{print}{SCData}(x, ...)
}
\arguments{
\item{x}{An SCData object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for SCData
}
