% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_sc.R
\name{print.ScSignature}
\alias{print.ScSignature}
\title{Print Method for ScSignature}
\usage{
\method{print}{ScSignature}(x, ...)
}
\arguments{
\item{x}{An ScSignature object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for ScSignature
}
