% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visium.R
\name{print.VisiumData}
\alias{print.VisiumData}
\title{Print Method for VisiumData}
\usage{
\method{print}{VisiumData}(x, ...)
}
\arguments{
\item{x}{A VisiumData object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for VisiumData
}
