% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_visium.R
\name{print.VisiumSignature}
\alias{print.VisiumSignature}
\title{Print Method for VisiumSignature}
\usage{
\method{print}{VisiumSignature}(x, ...)
}
\arguments{
\item{x}{A VisiumSignature object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for VisiumSignature
}
