% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{run_permutation_loop_internal}
\alias{run_permutation_loop_internal}
\title{Run Permutation Loop (Internal)}
\usage{
run_permutation_loop_internal(data_z, W, S0, n_perm, abs_observed, verbose)
}
\arguments{
\item{data_z}{Z-normalized expression matrix (genes x spots)}

\item{W}{Weight matrix (spots x spots)}

\item{S0}{Sum of weights}

\item{n_perm}{Number of permutations}

\item{abs_observed}{Absolute value of observed Moran's I matrix}

\item{verbose}{Print progress messages}
}
\value{
List with sum, sum_sq, count_extreme matrices
}
\description{
Internal function to run the permutation loop using pure R implementation.
Can be replaced with C++ implementation for better performance.
}
\keyword{internal}
