% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigdiscov-package.R
\docType{package}
\name{sigdiscov-package}
\alias{sigdiscov}
\alias{sigdiscov-package}
\title{sigdiscov: Spatial Correlation Analysis for Spatial Transcriptomics}
\description{
Compute pairwise Moran's I statistics and signed delta I signatures
for spatial transcriptomics data.
}
\section{Main Functions}{

\describe{
  \item{\code{\link{pairwise_moran}}}{Compute pairwise Moran's I between all genes}
  \item{\code{\link{compute_signed_delta_I}}}{Compute signed delta I signatures}
  \item{\code{\link{get_moran_curve}}}{Get I(r) curve for a single gene pair}
  \item{\code{\link{plot_moran_curve}}}{Visualize I(r) curve}
}
}

\section{Data Loading}{

\describe{
  \item{\code{\link{load_visium_data}}}{Load 10x Visium data from Space Ranger output}
  \item{\code{\link{filter_in_tissue}}}{Filter to in-tissue spots}
  \item{\code{\link{vst_transform}}}{VST normalization using sctransform}
  \item{\code{\link{run_pipeline}}}{Complete analysis pipeline}
}
}

\author{
\strong{Maintainer}: Peng Jiang \email{peng.jiang@nih.gov}

Authors:
\itemize{
  \item Seongyong Park \email{seongyong.park@nih.gov}
}

}
