% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_delta_I.R
\name{signed_delta_I}
\alias{signed_delta_I}
\title{Signed Delta I Signatures for Spatial Transcriptomics}
\description{
Functions for computing distance-dependent spatial correlation
signatures (signed delta I) between genes in spatial transcriptomics data.
}
