% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{single_gene_moran_cpp}
\alias{single_gene_moran_cpp}
\title{Univariate Moran's I for Each Gene}
\usage{
single_gene_moran_cpp(data_z, W, S0)
}
\arguments{
\item{data_z}{Z-normalized expression matrix (genes x spots)}

\item{W}{Weight matrix (spots x spots)}

\item{S0}{Sum of all weights}
}
\value{
Vector of Moran's I values
}
\description{
Computes univariate Moran's I for each gene (diagonal of pairwise matrix).
}
\keyword{internal}
