% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{singlecell}
\alias{singlecell}
\title{Single-Cell Spatial Transcriptomics Analysis}
\description{
Functions for computing cell-type-aware I_ND (Normalized Directional
Moran's I) signatures for single-cell spatial transcriptomics data (CosMx, Xenium,
MERFISH, etc.).
}
