% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{solve_diffusion}
\alias{solve_diffusion}
\title{Solve concentration field from sender cells}
\usage{
solve_diffusion(
  sender_positions,
  sender_expression,
  cell_positions,
  diffusion_config,
  n_receivers_density,
  position_dict = NULL,
  p_r = 1
)
}
\arguments{
\item{sender_positions}{Sender cell positions (n_senders x 2)}

\item{sender_expression}{Sender expression values}

\item{cell_positions}{All cell positions (n_cells x 2)}

\item{diffusion_config}{Diffusion configuration}

\item{n_receivers_density}{Receiver density for lambda calculation}

\item{position_dict}{Optional position dictionary for grouping}

\item{p_r}{Probability of responding}
}
\value{
List with concentrations and lambda
}
\description{
Solve concentration field from sender cells
}
