% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{stochastic_config}
\alias{stochastic_config}
\title{Create stochastic configuration}
\usage{
stochastic_config(
  p_sender_express = 0.9,
  p_receiver_respond_max = 1,
  hill_coefficient = 1,
  zero_inflate_factor = 0,
  zero_inflate_response = 0
)
}
\arguments{
\item{p_sender_express}{Probability sender is "on"}

\item{p_receiver_respond_max}{Max probability receiver responds}

\item{hill_coefficient}{Hill coefficient for response probability}

\item{zero_inflate_factor}{Fraction of non-sender cells with zero expression}

\item{zero_inflate_response}{Fraction of non-receiver cells with zero expression}
}
\value{
Stochastic configuration list
}
\description{
Create stochastic configuration
}
