% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomewide_analysis.R
\name{summary.GenomewideResult}
\alias{summary.GenomewideResult}
\title{Summary Method for GenomewideResult}
\usage{
\method{summary}{GenomewideResult}(object, ...)
}
\arguments{
\item{object}{GenomewideResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
Summary data frame
}
\description{
Summary Method for GenomewideResult
}
