% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{z_normalize_matrix_cpp}
\alias{z_normalize_matrix_cpp}
\title{Z-Normalize Expression Matrix}
\usage{
z_normalize_matrix_cpp(X)
}
\arguments{
\item{X}{Expression matrix (genes x spots)}
}
\value{
Z-normalized matrix
}
\description{
Z-normalizes each row (gene) of the expression matrix.
}
\keyword{internal}
