% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{load_moran_result}
\alias{load_moran_result}
\title{Load Moran's I Results}
\usage{
load_moran_result(input_file, gene_names = NULL)
}
\arguments{
\item{input_file}{Path to input file.}

\item{gene_names}{Optional character vector of gene names for row/column names.
If NULL, tries to read from a file with same name but .genes extension.}
}
\value{
A symmetric matrix of pairwise Moran's I values with row/column names.
}
\description{
Load pairwise Moran's I results from a file in lower triangular format.
}
